Update APDU parsing and secure command handling to support extended length fields
- Refactor `parse_apdu_case` to handle both short and extended APDU lengths, ensuring compatibility with larger data transmissions.
- Update `send_chained_data` for improved chunking logic and optimized segmentation of extended APDUs.
- Enhance secure messaging (`wrap_command`, `unwrap_command`) to support extended Lc and Le fields, ensuring proper encryption, padding, and error-handling.
- Add `OPGP_send_chained_APDU` for handling logical commands split across multiple APDUs, improving compatibility with long commands.
- Introduce tracing for wrapped commands and APDU responses to improve debugging and diagnostics.
- Improve memory management across APDU parsing and handling routines, minimizing leaks and optimizing allocations.

+++ gppcscconnectionplugin/src/gppcscconnectionplugin.c
@@ -410,8 +410,8 @@ OPGP_ERROR_STATUS OPGP_PL_send_APDU(OPGP_CARD_CONTEXT cardContext, OPGP_CARD_INF
 
 	DWORD result = SCARD_S_SUCCESS;
 	BYTE caseAPDU = 0;
-	BYTE lc = 0;
-	BYTE le = 0;
+	DWORD lc = 0;
+	DWORD le = 0;
 	BYTE la = 0;
 
 	DWORD offset = 0;
